add laravel policy for student and room

+++ app/Http/Controllers/StudentController.php
@@ -7,66 +7,63 @@ use Illuminate\Http\Request;
 use App\Service\UploadFileHandler;
 use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\StoreStudentRequest;
+use App\Repository\StudentRepository;
 
 class StudentController extends Controller
 {
 
     public $fileHandler;
     public function __construct(
+        StudentRepository $stuRepo,
         UploadFileHandler $fileHandler
-        ) {
+    ) {
+        $this->stuRepo = $stuRepo;
         $this->fileHandler = $fileHandler;
-
     }
     public function index(Request $request)
     {
         $user = Auth::user();
-        $students =  $user->students()->orderBy('created_at', 'Desc')
-        ->paginate(10);
-        
-        
+        $students =  $user->students()
+            ->orderBy('created_at', 'Desc')
+            ->paginate(10);
+
+
         return view('student.index', compact(
             'students'
         ));
     }
     public function create(Request $request)
     {
-        
+
         return view('student.create');
     }
     public function store(StoreStudentRequest $request)
     {
-        dd($request->all());
+        $this->stuRepo->storeStudent();
+        session()->flash('msg', 'Create Student Succeed');
+        return view('student.create');
     }
 
 
-    public function edit(Request $request)
-    {
-        return view('student.edit');
-    }
+
     public function update(Request $request, Student $student)
     {
-        $no = $request->get('no');
-        $name = $request->get('name');
-        $gender = $request->get('gender');
-        $file = $request->file('upload_img');
-        if($file != null){
-            $path = $this->fileHandler->saveStudentAvatar($file);
-            $student->photo = $path;
-        }
-        $student->no = $no;
-        $student->name = $name;
-        $student->gender = $gender;
-        $student->save();
+        $this->authorize('update', $student);
+        $this->stuRepo->updateStudent($student);
 
         session()->flash('msg', 'Update Action Succeed');
         return back();
     }
     public function destroy(Request $request, Student $student)
     {
+        $this->authorize('delete', $student);
         $this->fileHandler->deleteStudentAvatar($student->photo);
         $result = $student->delete();
         session()->flash('msg', 'Delete Action Succeed');
         return back();
+    }
+
+    public function downloadExcel(Request $request){
+        
     }
 }
